+++ tests/providers/microsoft/mssql/transfers/test_mssql_to_oracle.py
@@ -21,7 +21,7 @@ from unittest import mock
 
 from mock import MagicMock
 
-from airflow.providers.microsoft.mssql.transfers.mssql_to_oracle import MsSqlToOracleOperator
+from airflow.providers.microsoft.mssql.transfer.mssql_to_oracle import MsSqlToOracleOperator
 
 
 class TestMsSqlToOracle(unittest.TestCase):
